requirements.txt
streamlit>=1.32,<2.0
numpy>=1.24
opencv-python-headless>=4.8
Pillow>=10.0
